Added interval_mode parameter to servicechecks
In version 5, the Opsview REST API stores the check_interval in seconds,
whereas the old behavior was to record it in minutes.

opsview_servicecheck can make use of the 'interval_mode' parameter
to control how check_interval values are submitted and synched.

There are three modes: clever, minutes, and seconds.
Clever mode is the default.

Clever mode looks at the value of check_interval and uses minutes
if the value is less than 30 and seconds if it's greater than 30.

minutes mode treats all values as minutes.
seconds mode treats all values as seconds.